Add option for specifying an alement to append to
By default the calendar is inserted after the field it is being used on,
however you might need to append it to another object for positioning or
visibility purposes. This allows you to pass in a jQuery selector to use
as the target you want to append the calendar container to.