Implement positioning
Added the positionBelow() function that takes the element to position and
the target for it to go below and positions the element below it. This is
called when the calendar is created. Another method reposition() is
exposed on the plugin that is called on window resize and can be called
externally. This public method calls positionBelow using the the container
element and a selector that looks for the currently focused element as the
target, only does this if they calendar is open.